# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ES7-WebServices",
        "version": "1.0.1",
        "title": "ES7 WebService"
    },
    "basePath": "/api",
    "paths": {
        "/v1.0/EscuelaSabatica": {
            "get": {
                "tags": ["Escuela Sabatica"],
                "summary": "Obtener todos las Escuelas Sabaticas",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Lenguaje de Escuela Sabatica",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/config": {
            "get": {
                "tags": ["Escuela Sabatica"],
                "summary": "Obtener Escuela Sabatica publicada",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Lenguaje de Escuela Sabatica",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/{id}/": {
            "get": {
                "tags": ["Escuela Sabatica"],
                "summary": "Encontrar una Escuela Sabatica por su Identificador",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Identificador",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "404": {
                        "description": "Contacto no encotrado"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/create": {
            "post": {
                "tags": ["Escuela Sabatica"],
                "summary": "Crear una Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Escuela Sabatica",
                    "in": "body",
                    "description": "Escuela Sabatica a crear",
                    "required": true,
                    "schema": {
                            "type": "object",
                            "properties": {
                                "idEs": {
                                    "type": "integer"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "color": {
                                    "type": "string"
                                },
                                "quarterText": {
                                    "type": "string"
                                },
                                "quarter": {
                                    "type": "string"
                                },
                                "latestVersionName": {
                                    "type": "string"
                                },
                                "latestVersionCode": {
                                    "type": "integer"
                                },
                                "forceUpdate": {
                                    "type": "integer"
                                },
                                "googleUpdate": {
                                    "type": "integer"
                                },
                                "flexibleUpdate": {
                                    "type": "integer"
                                },
                                "helpVideos": {
                                    "type": "string"
                                },
                                "mission": {
                                    "type": "string"
                                },
                                "language": {
                                    "type": "string"
                                },
                                "egwHasAlert": {
                                    "type": "integer"
                                },
                                "appHasAlert": {
                                    "type": "integer"
                                },
                                "quarterNumber": {
                                    "type": "integer"
                                },
                                "published": {
                                    "type": "integer"
                                },
                                "frontPage": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                }
                            }
                        
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/inactive": {
            "patch": {
                "tags": ["Escuela Sabatica"],
                "summary": "Deshabilitar una Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Escuela Sabatica",
                    "in": "body",
                    "description": "Escuela Sabatica a Deshabilitar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "idEs": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/published": {
            "patch": {
                "tags": ["Escuela Sabatica"],
                "summary": "Publicar una Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Escuela Sabatica",
                    "in": "body",
                    "description": "Escuela Sabatica a Publicar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "idEs": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/EscuelaSabatica/update": {
            "put": {
                "tags": ["Escuela Sabatica"],
                "summary": "Editar una Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Escuela Sabatica",
                    "in": "body",
                    "description": "Escuela Sabatica a editar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "idEs": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "color": {
                                "type": "string"
                            },
                            "quarterText": {
                                "type": "string"
                            },
                            "quarter": {
                                "type": "string"
                            },
                            "latestVersionName": {
                                "type": "string"
                            },
                            "latestVersionCode": {
                                "type": "integer"
                            },
                            "forceUpdate": {
                                "type": "integer"
                            },
                            "googleUpdate": {
                                "type": "integer"
                            },
                            "flexibleUpdate": {
                                "type": "integer"
                            },
                            "helpVideos": {
                                "type": "string"
                            },
                            "mission": {
                                "type": "string"
                            },
                            "language": {
                                "type": "string"
                            },
                            "egwHasAlert": {
                                "type": "integer"
                            },
                            "appHasAlert": {
                                "type": "integer"
                            },
                            "quarterNumber": {
                                "type": "integer"
                            },
                            "published": {
                                "type": "integer"
                            },
                            "frontPage": {
                                    "type": "string"
                                },
                            "year": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
...
    }
}